refactor(services): split domain and infrastructure
