+++ backend/app/core/dependencies.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from fastapi import Depends
 
 from .config import get_settings, Settings
-from backend.services.openai_service import OpenAIService
-from backend.services.async_image_processor import AsyncImageProcessor
+from backend.services.infrastructure.openai_service import OpenAIService
+from backend.services.domain.async_image_processor import AsyncImageProcessor
 
 
 def get_openai_service(settings: Settings = Depends(get_settings)) -> OpenAIService:
@@ -16,7 +16,7 @@ def get_openai_service(settings: Settings = Depends(get_settings)) -> OpenAIServ
 
 def get_task_repository():
     """Return task manager module as a repository placeholder."""
-    from backend.services import task_manager  # local import to avoid circular
+    from backend.services.domain import task_manager  # local import to avoid circular
 
     return task_manager
 
@@ -31,6 +31,6 @@ def get_image_processor() -> AsyncImageProcessor:
 
 def get_process_image():
     """Provide the synchronous image processing function."""
-    from backend.services.image_processor import process_image
+    from backend.services.domain.image_processor import process_image
 
     return process_image
