+++ backend/services/__init__.py
@@ -1,0 +1,21 @@
+from .domain.async_image_processor import AsyncImageProcessor
+from .domain.image_processor import process_image
+from .domain.task_manager import (
+    TaskRecord,
+    create_task,
+    set_result,
+    set_error,
+    get_task,
+)
+from .infrastructure.openai_service import OpenAIService
+
+__all__ = [
+    "AsyncImageProcessor",
+    "process_image",
+    "TaskRecord",
+    "create_task",
+    "set_result",
+    "set_error",
+    "get_task",
+    "OpenAIService",
+]
