+++ backend/services/domain/async_image_processor.py
