+++ backend/services/domain/image_processor.py
